requirements.txt
flask==2.0.1
werkzeug==2.0.3
gunicorn==20.1.0
PyPDF2==3.0.1
openai>=1.0.0
python-dotenv==0.19.0
python-magic==0.4.24
scikit-learn>=0.24.2